+++ packages/core/src/persistent-storage/user-choices.ts
@@ -4,6 +4,7 @@ import { createLocalStorageInterface } from './local-storage-helpers';
 const USER_CHOICES_KEY = `${cssPrefix}-user-choices` as const;
 
 /**
+ * @public
  * Represents the user's choices for video and audio input devices,
  * as well as their username.
  */
@@ -33,20 +34,14 @@ export type LocalUserChoices = {
    * @defaultValue `''`
    */
   username: string;
-  /** @deprecated This property will be removed without replacement. */
-  e2ee: boolean;
-  /** @deprecated This property will be removed without replacement. */
-  sharedPassphrase: string;
 };
 
-const defaultUserChoices: LocalUserChoices = {
+export const defaultUserChoices: LocalUserChoices = {
   videoEnabled: true,
   audioEnabled: true,
   videoDeviceId: '',
   audioDeviceId: '',
   username: '',
-  e2ee: false,
-  sharedPassphrase: '',
 } as const;
 
 /**
@@ -72,9 +67,7 @@ export function saveUserChoices(
   if (preventSave === true) {
     return;
   }
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const { e2ee, sharedPassphrase, ...toSave } = userChoices;
-  save(toSave);
+  save(userChoices);
 }
 
 /**
@@ -98,8 +91,6 @@ export function loadUserChoices(
     videoDeviceId: defaults?.videoDeviceId ?? defaultUserChoices.videoDeviceId,
     audioDeviceId: defaults?.audioDeviceId ?? defaultUserChoices.audioDeviceId,
     username: defaults?.username ?? defaultUserChoices.username,
-    e2ee: defaults?.e2ee ?? defaultUserChoices.e2ee,
-    sharedPassphrase: defaults?.sharedPassphrase ?? defaultUserChoices.sharedPassphrase,
   };
 
   if (preventLoad) {