+++ packages/react/src/components/layout/CarouselLayout.tsx
@@ -21,12 +21,6 @@ export interface CarouselLayoutProps extends React.HTMLAttributes<HTMLMediaEleme
 }
 
 /**
- * @deprecated Renamed to [[CarouselLayout]]
- * @public
- */
-export const CarouselView = CarouselLayout;
-
-/**
  * The `CarouselLayout` component displays a list of tracks in a scroll container.
  * It will display as many tiles as possible and overflow the rest.
  * @remarks
@@ -48,8 +42,8 @@ export function CarouselLayout({ tracks, orientation, ...props }: CarouselLayout
   const carouselOrientation = orientation
     ? orientation
     : height >= width
-    ? 'vertical'
-    : 'horizontal';
+      ? 'vertical'
+      : 'horizontal';
 
   const tileSpan =
     carouselOrientation === 'vertical'